[solver] Add alternative implementation of solver that tries to prune the search space of conflicting routes.
The idea is to build a DAG that links how routes build on each other.
Then you can stop exploring the DAG once you hit a hexside that is
already claimed. This produces correct results, but ends up taking
many more iterations than the simpler version that just terminates
once the best-known revenue is out of reach. This is possibly because
the DAG version starts from the smallest routes and works its way up,
rather than the opposite. But its not completely clear to me why it
doesn't work better...